+++ ReferenceTests/src/tests/figures_and_makielayout.jl
@@ -133,6 +133,14 @@ end
     f
 end
 
+@reference_test "Log scale with manual ticks" begin
+    xs = range(10, 100; length=100)
+    f, ax, _ = lines(xs, xs)
+    ax.xticks = [10, 100]
+    ax.xscale = log10
+    f
+end
+
 @reference_test "Non-standard log tick formatting" begin
     xs = range(-10, 10; length=100)
     f, ax, _ = lines(xs, xs)